+++ superuser.py
@@ -15,6 +15,7 @@
 
 from discord.ext import commands
 import discord
+import git
 
 from required import *
 
@@ -64,7 +65,7 @@ class Superuser(commands.Cog):
         if not self.is_superuser(ctx.message.author.id):
             embed.title = "Permission Denied"
             embed.description = f'{ctx.message.author.name} is not the superuser. This incident will be reported.'
-            log(f'{ctx.message.author.name} is not the superuser. This incident will be reported.', self.log_file)
+            log(f'{ctx.message.author.name} attempted to restart the bot, but is not the superuser.', self.log_file)
             await ctx.send(embed=embed)
             return None
 
@@ -74,3 +75,33 @@ class Superuser(commands.Cog):
 
         subprocess.Popen(['python3', '__init__.py', '&'])
         await exit(0)
+
+    @commands.command()
+    async def pull(self, ctx):
+        '''
+        Pulls the latest commit from Git origin
+        '''
+
+        embed = discord.Embed(colour=random.randint(0, 16777215))
+
+        if not self.is_superuser(ctx.message.author.id):
+            embed.title = "Permission Denied"
+            embed.description = f'{ctx.message.author.name} is not the superuser. This incident will be reported.'
+            log(f'{ctx.message.author.name} attempted to pull the latest commit, but is not the superuser.', self.log_file)
+            await ctx.send(embed=embed)
+            return None
+
+        try:
+            g = git.cmd.Git("./")
+            g.pull()
+        except Exception as exception:
+            embed.title = "Git Pull Error"
+            embed.description = f'{exception}'
+            await ctx.send(embed=embed)
+            log(f'Git pull has failed with the following error: {exception}', self.log_file)
+
+        embed.title = "Successful Pull"
+        embed.description = "The latest commit was successfully pulled from origin."
+        await ctx.send(embed=embed)
+
+        log(f'The local copy of Torn-Bot has been updated by {ctx.message.author.name}.', self.log_file)