+++ CHANGELOG.md
@@ -20,6 +20,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
  - Requirement for version python 3.6 or greater
  - Configuration command that returns the current configuration
  - Random color to all embeds (issue #11)
+ - Pull command to pull the latest commit (Untested)
 ### Changed
  - Readme's TODO updated
  - Balance command to bal